+++ simpleGETproject/simpleget/src/main/java/com/alexdev/simpleget/services/UserService.java
@@ -8,9 +8,9 @@ public interface UserService {
 
     public List<User> getUsers();
 
-    public User getUserById(Integer userId);
+    public User getUserById(Long userId);
 
     public void saveUser(User newUser);
 
-    public void deleteUserById(Integer userId);
+    public void deleteUserById(Long userId);
 }
